MFC: A few more tweaks to support no-hardcoded-paths (1.28)

+++ direct/src/ffi/FFIInterrogateDatabase.py
@@ -715,7 +715,8 @@ class FFIInterrogateDatabase:
         # before we begin.
         for file in os.listdir(codeDir):
             pathname = os.path.join(codeDir, file)
-            os.unlink(pathname)
+            if not os.path.isdir(pathname):
+                os.unlink(pathname)
         
         # Import all the C++ modules
         for CModuleName in FFIConstants.CodeModuleNameList:
